+++ CHANGES.rst
@@ -7,12 +7,13 @@ Changes to be released in next version
 🙌 Improvements
  * MXKAttachment: Add support for all voice message types (vector-im/element-ios/issues/4094).
  * MXKRoomViewController, MXKRoomDataSource & MXKRoomInputToolbarView: Add methods to send videos as `AVAsset` objects (vector-im/element-ios/issues/4483).
+ * MXKSendReplyEventStringLocalizations: Added senderSentAVoiceMessage property
 
 🐛 Bugfix
  * 
 
 ⚠️ API Changes
- * 
+ * MXRoomDataSource: Added duration and sample parameters on the sendVoiceMessage method (vector-im/element-ios/issues/4090)
 
 🗣 Translations
  *