+++ CHANGES.rst
@@ -8,9 +8,10 @@ Changes to be released in next version
  * MXKAttachment: Add support for all voice message types (vector-im/element-ios/issues/4094).
  * MXKRoomViewController, MXKRoomDataSource & MXKRoomInputToolbarView: Add methods to send videos as `AVAsset` objects (vector-im/element-ios/issues/4483).
  * MXKSendReplyEventStringLocalizations: Added senderSentAVoiceMessage property
+ * MXKRoomDataSource: Retrying failed audio and voice messages
 
 🐛 Bugfix
- * 
+ * MXKRoomDataSource: Calling completion blocks for unsupported or failed message retries
 
 ⚠️ API Changes
  * MXRoomDataSource: Added duration and sample parameters on the sendVoiceMessage method (vector-im/element-ios/issues/4090)